Add 'kill yourself' button
You can now sacrifice if you get put into a bad state.
Also: the Save game will save the original board state to a file, not
the current state. This is because the load/save is for developing
levels, not saving progress in the current one. Perhaps I'll make it
optional as to which one you want to happpen somehow. Like
differentiating between 'save game' and 'export level definition' or
whatever