[Add cursor files]
I used the cursor files as guidelines when generating the code in this
commit. I one-shot it in cursor on Claude Sonnet 3.7 (non-thinking) with
the following prompt:

added files for context: README.md, board.js, game.js, utils.js
```text
Generate a class that can take a 'board' object and determine if the board is playable or not. the object is to retreive all of the crystals and get to the exit without falling in any holes. you push a block into a hole and it fills it in. You can push blocks through a one way door, so we can treat those as just 'walkable' so don't treat those as obstructions. someone may be able to fool this detector by cleverly arranging one way doors but that is a problem for another day. there should be a walkable path from the starting point to any required pushable blocks. there should be space for the player to actually push the block in the desired direction of the hole. separate these two path finding exercises into separate method calls, one for finding a walkable path between two squares (use A* path finding in 2 dimensions, short circuit returning true if any walkable path is found, no need to keep going once we determine there is one) Ensure there is a pushable path from each block to each hole that blocks a walkable path. this is slightly different from a walkable path because it will require two walkable squares, one for the player to stand in and one for the block to be pushed. keep this logic separate from any of the existing game logic but you may import and use any of it in this new class
```

`@README.md` `@game.js` `@board.js` `@utils.js`